dom/html/test/forms/test_input_datetime_tabindex.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1288591
-->
<head>
  <title>Test tabindex attribute for &lt;input type='time'&gt;</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1288591">Mozilla Bug 1288591</a>
<p id="display"></p>
<div id="content">
  <input id="time1" type="time" tabindex="0">
  <input id="time2" type="time" tabindex="-1">
  <input id="time3" type="time" tabindex="0">
</div>
<pre id="test">
<script type="application/javascript">

/**
 * Test for Bug 1288591.
 * This test checks whether date/time input types' tabindex attribute works
 * correctly.
 **/
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  test();
  SimpleTest.finish();
});

function test() {
  let time1 = document.getElementById("time1");
  let time2 = document.getElementById("time2");
  let time3 = document.getElementById("time3");

  time1.focus();
  is(document.activeElement, time1,
     "input element with tabindex=0 is focusable");

  // Advance to time1 minute field
  synthesizeKey("VK_TAB", {});
  is(document.activeElement, time1,
     "input element with tabindex=0 is tabbable");

  // Advance to time1 AM/PM field
  synthesizeKey("VK_TAB", {});
  is(document.activeElement, time1,
     "input element with tabindex=0 is tabbable");

  // Advance to next element
  synthesizeKey("VK_TAB", {});
  is(document.activeElement, time3,
     "input element with tabindex=-1 is not tabbable");

  time2.focus();
  is(document.activeElement, time2,
     "input element with tabindex=-1 is still focusable");

  // Changing the tabindex attribute dynamically.
  time3.setAttribute("tabindex", "-1");
  synthesizeKey("VK_TAB", {}); // need only one TAB since time2 is not tabbable
  isnot(document.activeElement, time3,
        "element with tabindex changed to -1 should not be tabbable");
}

</script>
</pre>
</body>
</html>
